Fix: moved the articles and quizzes pages to their correct place, updated version number
