+++ src/routes/index.tsx
@@ -4,6 +4,7 @@ import React from 'react';
 import { DefaultRoute, Routes } from './routes';
 import { LoadingScreen } from '../components/loadingscreen/LoadingScreen';
 import { useAppSelector } from '../lib/redux/Hooks';
+import { Articles } from '../screens/Articles';
 import CourseFinished from '../screens/Course/CourseFinished';
 import { Courses } from '../screens/Course/Courses';
 import Stage from '../screens/Course/Stage';
@@ -11,6 +12,7 @@ import StageOverview from '../screens/Course/StageOverview';
 import { HomeScreen } from '../screens/Home';
 import { PromptLibrary } from '../screens/PromptLibrary/PromptLibrary';
 import { PromptView } from '../screens/PromptLibrary/PromptView';
+import { Quizzes } from '../screens/Quizzes';
 import { SettingsScreen } from '../screens/Settings';
 import { StudyScreen } from '../screens/Study';
 import { CaseStudyInfo } from '../screens/Usecase/CaseStudyInfo';
@@ -18,8 +20,6 @@ import { UsecaseScreen } from '../screens/Usecase/Usecase';
 import BackgroundCollectForm from '../screens/UserBackground/BackgroundCollectForm';
 import { BackgroundInfo } from '../screens/UserBackground/BackgroundInfo';
 import { WTRScreen } from '../screens/WTR';
-import { Articles } from '../screens/WTR/Articles';
-import { Quizzes } from '../screens/WTR/Quizzes';
 import { WTRContentScreen } from '../screens/WTR/WTRContent';
 
 const Stack = createNativeStackNavigator();
